requirements.txt
joblib==1.3.2
numpy==1.26.4
pillow==10.2.0
scikit-learn==1.4.0
scipy==1.12.0
threadpoolctl==3.3.0